+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.0.2
+- Fix Issue #1: Custom functions are not being attached to po if it is sub-element of 'makeParentPo'
+- Fix Issue #2: makeListPo elements cannot be nested within makeParentPo
+
 # v1.0.1
 - Allow custom functions to be included in page object via els `fns` property 
 - Update README to include documentation for ngpo functions